+++ SP's_Marina_Mooring_Rates/SP_UPDATE_MARINA_MOORING_RATE.sql
@@ -6,7 +6,7 @@ CREATE PROCEDURE `SP_UPDATE_MARINA_MOORING_RATE`
     _description varchar(300),
     _ft_min SMALLINT,
     _ft_max SMALLINT,
-    _price FLOAT(16,8)
+    _price DECIMAL(12,4)
 )
 
 BEGIN
